requirements.txt
Django==1.8.5
djangorestframework==3.6.3
whitenoise==2.0
